refactor: move makeRecipe to its own file
