+++ src/game-logic/reducers/makeRecipe.test.js
@@ -1,0 +1,38 @@
+import { sampleRecipe1 } from '../../data/recipes'
+
+import { makeRecipe } from './makeRecipe'
+
+jest.mock('../../data/recipes')
+
+describe('makeRecipe', () => {
+  describe('there are insufficient ingredients for recipe', () => {
+    test('the recipe is not made', () => {
+      const { inventory } = makeRecipe(
+        {
+          inventory: [{ id: 'sample-item-1', quantity: 1 }],
+          inventoryLimit: -1,
+        },
+        sampleRecipe1
+      )
+
+      expect(inventory).toEqual([{ id: 'sample-item-1', quantity: 1 }])
+    })
+  })
+
+  describe('there are sufficient ingredients for recipe', () => {
+    test('consumes ingredients and adds recipe item to inventory', () => {
+      const { inventory } = makeRecipe(
+        {
+          inventory: [{ id: 'sample-item-1', quantity: 3 }],
+          inventoryLimit: -1,
+        },
+        sampleRecipe1
+      )
+
+      expect(inventory).toEqual([
+        { id: 'sample-item-1', quantity: 1 },
+        { id: 'sample-recipe-1', quantity: 1 },
+      ])
+    })
+  })
+})
